requirements.txt
python-telegram-bot==20.3
psycopg2-binary
requests
beautifulsoup4
python-dotenv
nest_asyncio
pytest
pytest-asyncio
